requirements.txt
asgiref==3.6.0
Django==3.2.4
django-json-widget==1.1.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
future==0.18.2
psycopg2-binary==2.9.5
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
